Added a float to the input variable

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_income = int(input('Enter your monthly income: '))
-monthly_expenses = int(input('Enter your monthly expenses: '))
+monthly_income = float(input('Enter your monthly income: '))
+monthly_expenses = float(input('Enter your monthly expenses: '))
 monthly_savings = int(float(monthly_income) - float(monthly_expenses))
 annual_interest_rate = 0.05
 projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * annual_interest_rate))
